Updated types and indexes

+++ script/bybit_ta_spot_tables.sql
@@ -8,8 +8,8 @@ create TABLE IF NOT EXISTS crypto_scout.bybit_ta_spot_public_trade (
     id BIGSERIAL,
     symbol TEXT NOT NULL,
     trade_time TIMESTAMP WITH TIME ZONE NOT NULL,
-    price NUMERIC(20, 8) NOT NULL,
-    size NUMERIC(20, 8) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    size DOUBLE PRECISION NOT NULL,
     taker_side TEXT NOT NULL CHECK (taker_side IN ('Buy','Sell')),
     is_block_trade BOOLEAN NOT NULL,
     is_rpi BOOLEAN NOT NULL,
@@ -39,8 +39,8 @@ create TABLE IF NOT EXISTS crypto_scout.bybit_ta_spot_order_book_1 (
     symbol TEXT NOT NULL,
     engine_time TIMESTAMP WITH TIME ZONE NOT NULL,
     side TEXT NOT NULL CHECK (side IN ('bid','ask')),
-    price NUMERIC(20, 8) NOT NULL,
-    size NUMERIC(20, 8) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    size DOUBLE PRECISION NOT NULL,
     CONSTRAINT bybit_ta_spot_order_book_1_pkey PRIMARY KEY (id, engine_time)
 );
 alter table crypto_scout.bybit_ta_spot_order_book_1 OWNER TO crypto_scout_db;
@@ -54,8 +54,8 @@ create TABLE IF NOT EXISTS crypto_scout.bybit_ta_spot_order_book_50 (
     symbol TEXT NOT NULL,
     engine_time TIMESTAMP WITH TIME ZONE NOT NULL,
     side TEXT NOT NULL CHECK (side IN ('bid','ask')),
-    price NUMERIC(20, 8) NOT NULL,
-    size NUMERIC(20, 8) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    size DOUBLE PRECISION NOT NULL,
     CONSTRAINT bybit_ta_spot_order_book_50_pkey PRIMARY KEY (id, engine_time)
 );
 alter table crypto_scout.bybit_ta_spot_order_book_50 OWNER TO crypto_scout_db;
@@ -69,8 +69,8 @@ create TABLE IF NOT EXISTS crypto_scout.bybit_ta_spot_order_book_200 (
     symbol TEXT NOT NULL,
     engine_time TIMESTAMP WITH TIME ZONE NOT NULL,
     side TEXT NOT NULL CHECK (side IN ('bid','ask')),
-    price NUMERIC(20, 8) NOT NULL,
-    size NUMERIC(20, 8) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    size DOUBLE PRECISION NOT NULL,
     CONSTRAINT bybit_ta_spot_order_book_200_pkey PRIMARY KEY (id, engine_time)
 );
 alter table crypto_scout.bybit_ta_spot_order_book_200 OWNER TO crypto_scout_db;
@@ -84,8 +84,8 @@ create TABLE IF NOT EXISTS crypto_scout.bybit_ta_spot_order_book_1000 (
     symbol TEXT NOT NULL,
     engine_time TIMESTAMP WITH TIME ZONE NOT NULL,
     side TEXT NOT NULL CHECK (side IN ('bid','ask')),
-    price NUMERIC(20, 8) NOT NULL,
-    size NUMERIC(20, 8) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
+    size DOUBLE PRECISION NOT NULL,
     CONSTRAINT bybit_ta_spot_order_book_1000_pkey PRIMARY KEY (id, engine_time)
 );
 alter table crypto_scout.bybit_ta_spot_order_book_1000 OWNER TO crypto_scout_db;
